+++ tools/njsq/wrapper.cpp
@@ -371,7 +371,9 @@ qRB
 
 	Caller.Init( Info );
 
-	n4allw::GetLauncher().Launch( n4allw::GetFunction( Index->Uint32Value()), Caller );
+	n4allw::Launch( Index->Uint32Value(), Caller );
+
+//	n4allw::GetLauncher().Launch( n4allw::GetFunction( Index->Uint32Value() ), Caller );
 
 	if ( sclerror::IsErrorPending() )
 		qRAbort();	// To force the handling of a pending error.