+++ tools/njsq/CHANGELOG.md
@@ -1,5 +1,8 @@
 # *NJSq* changelog
 
+2017-10-10 :
+- Using new `n4allw::Launch(...)` function.
+
 2017-09-17 :
 - Upgrading.